oauth update

+++ src/main/java/com/aot/groups/repositories/GroupRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aot.groups.models.Group;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 	
 	Optional<Group> findByName(String name);
